// src/routes/users/users.schema.ts
export const UserStatusEnum = z.enum([
  "ACTIVE",
  "INACTIVE",
  "SUSPENDED",
  "PENDING",
]);
export const UserRoleEnum = z.enum(["ADMIN", "USER"]);

// Base User schema for validation
export const UserSchema = z.object({
  id: z.string(),
  clerkUserId: z.string(),
  email: z.string().email("Invalid email address"),
  firstName: z.string().min(1, "First name is required"),
  lastName: z.string().min(1, "Last name is required"),
  name: z.string().min(1, "Name is required"),
  phone: z.string().optional().nullable(),
  emailVerified: z.date().optional().nullable(),
  image: z.string().url("Image must be a valid URL").optional().nullable(),
  role: UserRoleEnum.default("USER"),
  password: z.string().optional().nullable(),
  status: UserStatusEnum.default("ACTIVE"),
  isVerified: z.boolean().default(true),
  token: z.string().optional().nullable(),
  resetExpiry: z.date().optional().nullable(),
  createdAt: z.date().optional(),
  updatedAt: z.date().optional(),
  roleId: z.string().optional().nullable(),
});